% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_position_quantile.R
\name{adjusted_quantile}
\alias{adjusted_quantile}
\title{Compute an adjusted quantile that handles rank ties}
\usage{
adjusted_quantile(min_quantile, max_quantile)
}
\arguments{
\item{min_quantile}{quantile of an observation, computed by
taking rank to be the smallest among
tied observations}

\item{max_quantile}{quantile of an observation, computed by
taking rank to be the largest among
tied observations}
}
\value{
a quantile computed as described above
}
\description{
Compares the minimum and maximum rank of an observation and
returns the quantile closest to the median.
}
