% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_dist_to_array.R
\name{apply_dist_to_array}
\alias{apply_dist_to_array}
\title{Apply a function to each pixel of (2,3)D array, passing fit parameters as reference}
\usage{
apply_dist_to_array(dist, obs, fn, when.dist.undefined = NA)
}
\arguments{
\item{dist}{a (2,3)D array of distribution fit parameters}

\item{obs}{a (2,3)D array of observation parameters}

\item{fn}{A function to apply the distribution to the observations,
when the distribution is defined.  Should have signature
\code{function(values, dist_params)}. The function should return
a vector of the same length for all pixels.}

\item{when.dist.undefined}{A value to use when the distribution is
undefined. Length must match the return
value of \code{fn}.}
}
\value{
a (2,3)D array of values returned by \code{fn}
}
\description{
Apply a function to each pixel of (2,3)D array, passing fit parameters as reference
}
