% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_apply.R
\name{array_apply}
\alias{array_apply}
\title{Apply a function to each pixel of a stack of matrices}
\usage{
array_apply(arr, fun, ...)
}
\arguments{
\item{arr}{a three-dimensional array}

\item{fun}{a function to apply at each row and column of the input
array. The function will be called with a vector of the
z-values as a single argument.}

\item{...}{additional arguments to be passed to \code{fun}}
}
\value{
an array containing the values returned by \code{fun} at
        each pixel
}
\description{
Apply a function to each pixel of a stack of matrices
}
