% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_position_quantile.R
\name{cdf_plotting_position}
\alias{cdf_plotting_position}
\title{Compute a comulative probablility for an observation, given an ordered list of historical observations}
\usage{
cdf_plotting_position(obs, ranked_historical_obs, plotting_position_fn)
}
\arguments{
\item{obs}{observation for which a probability should be computed}

\item{ranked_historical_obs}{an array of historical observations, stored as sorted
values along the 3rd dimension}

\item{plotting_position_fn}{a function that computes a probability given a rank
and number of *historical* observations as arguments}
}
\value{
cumulative probability for each values of \code{obs}
}
\description{
Compute a comulative probablility for an observation, given an ordered list of historical observations
}
