% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdfgev}
\alias{cdfgev}
\title{Cumulative distribution function for the generalized extreme value (GEV) distribution}
\usage{
cdfgev(x, location, scale, shape)
}
\arguments{
\item{x}{vector of quantiles}

\item{location}{vector of location parameters}

\item{scale}{vector of scale}

\item{shape}{vector of shape parameters}
}
\description{
Cumulative distribution function for the generalized extreme value (GEV) distribution
}
