% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cdfpe3}
\alias{cdfpe3}
\title{Cumulative distribution function for the Pearson Type-III distribution}
\usage{
cdfpe3(x, location, scale, shape)
}
\arguments{
\item{x}{vector of quantiles}

\item{location}{vector of location parameters}

\item{scale}{vector of scale}

\item{shape}{vector of shape parameters}
}
\value{
probability assocated with each quantile \code{x}
}
\description{
The provided distribution parameters (\code{location}, \code{scale},
\code{shape}) must have either a length of \code{1}, or the same length
as \code{x}.
}
