% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_cdf.R
\name{find_cdf}
\alias{find_cdf}
\title{Lookup a CDF by distribution name}
\usage{
find_cdf(distribution)
}
\arguments{
\item{distribution}{name of a statistical distribution}
}
\value{
cumulative distribution function, or NULL if no
}
\description{
Lookup a CDF by distribution name
}
