% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_qua.R
\name{find_qua}
\alias{find_qua}
\title{Lookup a quantile function (inverse CDF) by distribution name}
\usage{
find_qua(distribution)
}
\arguments{
\item{distribution}{name of a statistical distribution}
}
\value{
quantile function, or NULL if not found
}
\description{
Lookup a quantile function (inverse CDF) by distribution name
}
