% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stat.R
\name{find_stat}
\alias{find_stat}
\title{Get a function to compute a named statistic over a vector of observations}
\usage{
find_stat(name)
}
\arguments{
\item{name}{name of statistic. Supported statistics are:
\describe{
\item{min}{minimum defined value}
\item{max}{maximum defined value}
\item{ave}{mean defined value}
\item{median}{median defined value}
\item{qXX}{quantile \code{XX} of defined values}
}}
}
\description{
Get a function to compute a named statistic over a vector of observations
}
