% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cell_distributions.R
\name{fit_cell_distributions}
\alias{fit_cell_distributions}
\title{Fit cell-wise statistical distributions using L-moments}
\usage{
fit_cell_distributions(
  distribution,
  arr,
  nmin.unique = 10,
  nmin.defined = 10,
  zero.scale.to.na = TRUE,
  log.errors = NULL
)
}
\arguments{
\item{distribution}{The name of the distribution to fit (e.g., "gev", "pe3").
Alternatively, \code{distribution} can be set to "nonparametric"
in which case all pixelwise values will be returned in
sorted order.}

\item{arr}{A 3D array representing multiple observations of a variable
in each 2D cell}

\item{nmin.unique}{Minimum number of unique values required to perform
a parametric distribution fit}

\item{nmin.defined}{Minimum number of defined values required to perform
a parametric distribution fit}

\item{zero.scale.to.na}{If TRUE, fit will be discarded (set to NA) if
the parametric distribution scale parameter is
computed to be zero.}

\item{log.errors}{Function that will be called with any errors reported
by distribution-fitting function. If \code{log.errors}
is not specified, errors will be written to stdout.}
}
\value{
A 3D array containing the fitted parameters of a parametric distribution at
        each pixel, or the the sorted values of \code{arr} when
        \code{distribution = 'nonparametric'}.
}
\description{
If the observed data in a cell contains an insufficient number of
defined values, then all distribution parameters at that pixel will
be undefined.
}
\details{
If the observed data at a pixel contains a a sufficient number of
defined values, but an insufficient number of unique values, then the
location parameter at that pixel will be set to the median of the
defined values and the other parameters will remain undefined.

If a distribution cannot be fit, then the returned parameters will
be NA.  (TODO: Describe cases where this may happen.)
}
