% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_backend_exists.R
\name{parallel_backend_exists}
\alias{parallel_backend_exists}
\title{Test whether the parallel package has a default cluster}
\usage{
parallel_backend_exists()
}
\value{
TRUE if a default cluster is registered
}
\description{
Test whether the parallel package has a default cluster
}
