% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_position_quantile.R
\name{plotting_position_tukey}
\alias{plotting_position_tukey}
\title{Compute the Tukey plotting position}
\usage{
plotting_position_tukey(rank, n_obs)
}
\arguments{
\item{rank}{rank of an observation}

\item{n_obs}{number of observations, *excluding* the observation
whose plotting position is being computed}
}
\value{
cumulative probability
}
\description{
Tukey plotting position is computed using the formulation in Wilks (2011)
equation 7.43.
}
