% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quagev}
\alias{quagev}
\title{Quantile function for the generalized extreme value (GEV) distribution}
\usage{
quagev(x, location, scale, shape)
}
\arguments{
\item{x}{vector of probabilities}

\item{location}{vector of location parameters}

\item{scale}{vector of scale}

\item{shape}{vector of shape parameters}
}
\description{
Quantile function for the generalized extreme value (GEV) distribution
}
