% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile2correct.R
\name{quantile2correct}
\alias{quantile2correct}
\title{quantile2correct}
\usage{
quantile2correct(quant, obs_fit)
}
\arguments{
\item{quant}{matrix with quantiles of the forecast to correct}

\item{obs_fit}{3D array with parameters for the observed value distribution}
}
\value{
matrix with a corrected forecast
}
\description{
Given the quantile of a forecast and a distribution of observations,
produce a corrected forecast by identifying the observed value whose
quantile corresponds to the quantile of the forecast.
}
