% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile2rp.R
\name{quantile2rp}
\alias{quantile2rp}
\title{Convert quantiles into return periods}
\usage{
quantile2rp(q)
}
\arguments{
\item{q}{vector or matrix of quantiles}
}
\value{
return period for each quantile
}
\description{
Convert quantiles into return periods
}
