% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quape3}
\alias{quape3}
\title{Quantile function for the Pearson Type-III distribution}
\usage{
quape3(x, location, scale, shape)
}
\arguments{
\item{x}{vector of probabilities}

\item{location}{vector of location parameters}

\item{scale}{vector of scale}

\item{shape}{vector of shape parameters}
}
\value{
quantiles associated with each probability \code{x}
}
\description{
The provided distribution parameters (\code{location}, \code{scale},
\code{shape}) must have either a length of \code{1}, or the same length
as \code{x}.
}
