% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2quantile.R
\name{raw2quantile}
\alias{raw2quantile}
\title{Calculate the quantile of a raw forecast}
\usage{
raw2quantile(forecast, retro_fit)
}
\arguments{
\item{forecast}{a matrix of forecasted values}

\item{retro_fit}{a 3D array of GEV fit parameters from retrospective forecasts}
}
\value{
a matrix of quantiles for each observation
}
\description{
Calculate the quantile of a raw forecast relative to retrospective forecasts
}
