% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp2quantile.R
\name{rp2quantile}
\alias{rp2quantile}
\title{Convert return periods into quantiles}
\usage{
rp2quantile(rp)
}
\arguments{
\item{rp}{vector or matrix of return periods}
}
\value{
quantile corresponding to each rp
}
\description{
Convert return periods into quantiles
}
