% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp2sa.R
\name{rp2sa}
\alias{rp2sa}
\title{Convert return period into standardized anomaly}
\usage{
rp2sa(rp)
}
\arguments{
\item{rp}{vector or matrix of return periods}
}
\value{
standardized anomalies
}
\description{
Convert return period into standardized anomaly
}
