% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa2rp.R
\name{sa2rp}
\alias{sa2rp}
\title{Convert standardized anomaly into a return period}
\usage{
sa2rp(sa, min.rp = -1000, max.rp = 1000)
}
\arguments{
\item{sa}{vector or matrix of standardized anomalies}

\item{min.rp}{minimum value for clamping return period}

\item{max.rp}{maximum value for clamping return period}
}
\value{
return period
}
\description{
Convert standardized anomaly into a return period
}
