% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_frac_defined}
\alias{stack_frac_defined}
\title{Compute the fraction of defined elements for each row and col in a 3D array}
\usage{
stack_frac_defined(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the computed fraction for each [row, col, ]
}
\description{
Compute the fraction of defined elements for each row and col in a 3D array
}
