% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_max_rank}
\alias{stack_max_rank}
\title{Compute the rank of each element in a matrix, returning the maximum in case of ties}
\usage{
stack_max_rank(x, obs)
}
\arguments{
\item{x}{a matrix of values to rank}

\item{obs}{a 3D array of observations against which each value in x should be ranked}
}
\value{
the rank of \code{x} after it is added to \code{obs}, for each (i, j) in \code{x}
}
\description{
Compute the rank of each element in a matrix, returning the maximum in case of ties
}
