% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_mean}
\alias{stack_mean}
\title{Compute the mean defined element for each row and col in a 3D array}
\usage{
stack_mean(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the mean value for each [row, col, ]
}
\description{
Compute the mean defined element for each row and col in a 3D array
}
