% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_median}
\alias{stack_median}
\title{Compute the median of defined elements for each row and col in a 3D array}
\usage{
stack_median(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the median for each [row, col, ]
}
\description{
Compute the median of defined elements for each row and col in a 3D array
}
