% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_min}
\alias{stack_min}
\title{Compute the minimum defined element for each row and col in a 3D array}
\usage{
stack_min(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the minimum value for each [row, col, ]
}
\description{
Compute the minimum defined element for each row and col in a 3D array
}
