% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_num_defined}
\alias{stack_num_defined}
\title{Compute the number of defined elements for each row and col in a 3D array}
\usage{
stack_num_defined(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the computed count for each [row, col, ]
}
\description{
Compute the number of defined elements for each row and col in a 3D array
}
