% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_quantile}
\alias{stack_quantile}
\title{Compute a given quantile of defined elements for each row and col in a 3D array}
\usage{
stack_quantile(v, q)
}
\arguments{
\item{v}{3D array that may contain NA values}

\item{q}{a quantile to compute, q within [0,1]}
}
\value{
a matrix with the specified quantile for each [row, col, ]
}
\description{
Compute a given quantile of defined elements for each row and col in a 3D array
}
