% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_sort}
\alias{stack_sort}
\title{Sort each slice [i, j, ] in an array}
\usage{
stack_sort(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the median for each [row, col, ]
}
\description{
Sort each slice [i, j, ] in an array
}
