% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_sum}
\alias{stack_sum}
\title{Compute the sum of defined elements for each row and col in a 3D array}
\usage{
stack_sum(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the sum for each [row, col, ]
}
\description{
Compute the sum of defined elements for each row and col in a 3D array
}
