% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_weighted_mean}
\alias{stack_weighted_mean}
\title{Compute a weighted mean of defined elements for each row and col in a 3D array}
\usage{
stack_weighted_mean(v, w)
}
\arguments{
\item{v}{a 3D array that may contain NA values}

\item{w}{a 2D vector of weights, having the same length as the third dimension of \code{v}}
}
\value{
a matrix with the specified quantile for each [row, col, ]
}
\description{
Compute a weighted mean of defined elements for each row and col in a 3D array
}
