% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_weighted_quantile}
\alias{stack_weighted_quantile}
\title{Compute a given weighted quantile of defined elements for each row and col in a 3D array}
\usage{
stack_weighted_quantile(v, w, q)
}
\arguments{
\item{v}{a 3D array that may contain NA values}

\item{w}{a 2D vector of weights, having the same length as the third dimension of \code{v}}

\item{q}{a quantile to compute, q within [0, 1]}
}
\value{
a matrix with the specified quantile for each [row, col, ]
}
\description{
Compute a given weighted quantile of defined elements for each row and col in a 3D array
}
