% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_which_max}
\alias{stack_which_max}
\title{Compute the index of the maximum defined element for each row and col in a 3D array}
\usage{
stack_which_max(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the maximum index for each [row, col, ]
}
\description{
Compute the index of the maximum defined element for each row and col in a 3D array
}
