% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stack_which_min}
\alias{stack_which_min}
\title{Compute the index of the minimum defined element for each row and col in a 3D array}
\usage{
stack_which_min(v)
}
\arguments{
\item{v}{3D array that may contain NA values}
}
\value{
a matrix with the minimum index for each [row, col, ]
}
\description{
Returned value is indexed beginning at 1
}
