% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_anomaly.R
\name{standard_anomaly}
\alias{standard_anomaly}
\title{Compute the standardized anomaly associated with an observation}
\usage{
standard_anomaly(
  distribution,
  dist_params,
  obs,
  plotting_position = "tukey",
  min.sa = -100,
  max.sa = 100
)
}
\arguments{
\item{distribution}{name of distribution used for \code{dist_params}}

\item{dist_params}{3D arary of distribution parameters}

\item{obs}{observed value for which a standard anomaly should be computed}

\item{plotting_position}{when \code{distribution} is \code{nonparametric},
the name of the plotting position to use (e.g., 'tukey')}

\item{min.sa}{minimum value for clamping computed standardized anomaly}

\item{max.sa}{maximum value for clamping computed standardized anomaly}
}
\value{
computed standardized anomaly
}
\description{
Compute the standardized anomaly associated with an observation
}
