% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{substitute}
\alias{substitute}
\title{Substitute specifed values in a vector with replacements}
\usage{
substitute(vals, subs)
}
\arguments{
\item{vals}{A vector in which some values should be replaced}

\item{subs}{A vector containing a sequence of replacement values,
in the form val_1, replacement_1, val_2, replacement_2, ...}
}
\description{
Substitute specifed values in a vector with replacements
}
