% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wsim_quantile}
\alias{wsim_quantile}
\title{Compute a sample quantile of a given vector. NA values are ignored.}
\usage{
wsim_quantile(v, q)
}
\arguments{
\item{v}{a vector of numeric values, possibly contining NAs}

\item{q}{quantile to compute}
}
\value{
sample quantile \code{q} of \code{v}
}
\description{
This function is provided for improved performance over the built-in
\code{quantile} function. Interpolation is performed using the default
method 7 specified in documentation for \code{quantile}.
}
